backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.0